+++ app/src/main/java/hr/dtakac/prognoza/uimodel/cell/HourUiModel.kt
@@ -6,15 +6,15 @@ import java.time.ZonedDateTime
 
 data class HourUiModel(
     val id: String,
-    val temperature: Float?,
-    val feelsLike: Float?,
-    val precipitationAmount: Float?,
-    val windSpeed: Float?,
+    val temperature: Double?,
+    val feelsLike: Double?,
+    val precipitationAmount: Double?,
+    val windSpeed: Double?,
     val windFromCompassDirection: Int?,
     val weatherDescription: WeatherDescription?,
     val time: ZonedDateTime,
-    val relativeHumidity: Float?,
-    val airPressureAtSeaLevel: Float?,
+    val relativeHumidity: Double?,
+    val airPressureAtSeaLevel: Double?,
     val displayDataInUnit: MeasurementUnit,
     var isExpanded: Boolean = false
 )
